refactor(types): extract hotkey/shortcut/capability types into types/hotkey.rs — behavior unchanged, 446 tests green
Pure move of 15 types (ShortcutBinding, QaHotkeyBinding, ComboBinding, HotkeyTrigger/
Mode/AdapterKind, HotkeyKey/Binding, HotkeyCapability/InstallError/Status, WindowsIme*,
PlatformCapabilities, HotkeyStatusState) + their 13 impls + 4 private helper fns out of
types.rs into a child module. Re-exported via 'pub use hotkey::*;' so every
'crate::types::Hotkey*' import across the codebase keeps resolving — serde contract and
public API byte-identical. types.rs 2900→2300 lines.

cargo check (+--tests): 0 errors, 87 warnings (unchanged) · clippy: 0 errors
cargo test --lib: 446 passed · tsc / npm run build: green
