+++ openless-all/app/src-tauri/src/types/hotkey.rs
@@ -1,0 +1,611 @@
+//! Hotkey / shortcut / platform-capability types extracted from `types.rs`
+//! (behavior-preserving move).
+//!
+//! Pure type definitions + impls; re-exported from `types` via `pub use hotkey::*;`
+//! so existing `crate::types::HotkeyBinding` (etc.) imports keep resolving.
+
+use super::*;
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ShortcutBinding {
+    pub primary: String,
+    pub modifiers: Vec<String>,
+}
+
+impl ShortcutBinding {
+    pub fn default_qa() -> Self {
+        #[cfg(target_os = "macos")]
+        {
+            Self {
+                primary: ";".into(),
+                modifiers: vec!["cmd".into(), "shift".into()],
+            }
+        }
+        #[cfg(not(target_os = "macos"))]
+        {
+            Self {
+                primary: ";".into(),
+                modifiers: vec!["ctrl".into(), "shift".into()],
+            }
+        }
+    }
+
+    pub fn display_label(&self) -> String {
+        let mut parts: Vec<String> = Vec::new();
+        let modifier_order = ["cmd", "ctrl", "alt", "shift", "super"];
+        for tag in modifier_order {
+            if self.modifiers.iter().any(|m| m.eq_ignore_ascii_case(tag)) {
+                parts.push(modifier_display(tag).to_string());
+            }
+        }
+        parts.push(display_primary(&self.primary));
+        parts.join("+")
+    }
+}
+
+/// 划词语音问答的全局快捷键绑定。原生名字符串：
+/// - `primary`：主键（如 `";"`、`"."`、`"A"`、`"F1"`）。
+/// - `modifiers`：修饰键集合，元素来自 `{"cmd","ctrl","alt","shift","super"}`。
+///   小写名简单序列化即可，前端 / 后端解析时统一 lowercase。
+///
+/// 默认 `Cmd+Shift+;` (macOS) / `Ctrl+Shift+;` (Windows)。
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct QaHotkeyBinding {
+    pub primary: String,
+    pub modifiers: Vec<String>,
+}
+
+impl Default for QaHotkeyBinding {
+    fn default() -> Self {
+        #[cfg(target_os = "macos")]
+        {
+            Self {
+                primary: ";".into(),
+                modifiers: vec!["cmd".into(), "shift".into()],
+            }
+        }
+        #[cfg(not(target_os = "macos"))]
+        {
+            Self {
+                primary: ";".into(),
+                modifiers: vec!["ctrl".into(), "shift".into()],
+            }
+        }
+    }
+}
+
+impl QaHotkeyBinding {
+    /// 渲染成给前端展示的可读标签。
+    /// 顺序与人类阅读习惯一致：`Cmd+Shift+;`、`Ctrl+Alt+Shift+.`。
+    pub fn display_label(&self) -> String {
+        let mut parts: Vec<String> = Vec::new();
+        // 固定输出顺序：Ctrl/Cmd → Alt/Option → Shift → Super
+        let modifier_order = ["cmd", "ctrl", "alt", "shift", "super"];
+        for tag in modifier_order {
+            if self.modifiers.iter().any(|m| m.eq_ignore_ascii_case(tag)) {
+                parts.push(modifier_display(tag).to_string());
+            }
+        }
+        let key_label = display_primary(&self.primary);
+        parts.push(key_label);
+        parts.join("+")
+    }
+}
+
+/// 录音快捷键的自定义组合键绑定。结构与 `QaHotkeyBinding` 相同：
+/// - `primary`：主键（如 `"D"`、`"Space"`、`"F1"`）。
+/// - `modifiers`：修饰键集合，元素来自 `{"cmd","ctrl","alt","shift","super"}`。
+///
+/// 当 `HotkeyBinding.trigger == Custom` 时，coordinator 用 `global-hotkey` crate
+/// 注册此组合键，而非 modifier-only 的 CGEventTap / WH_KEYBOARD_LL。
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ComboBinding {
+    pub primary: String,
+    pub modifiers: Vec<String>,
+}
+
+impl ComboBinding {
+    /// 渲染成给前端展示的可读标签。复用 QaHotkeyBinding 的格式化逻辑。
+    pub fn display_label(&self) -> String {
+        let qa = QaHotkeyBinding {
+            primary: self.primary.clone(),
+            modifiers: self.modifiers.clone(),
+        };
+        qa.display_label()
+    }
+}
+
+pub(super) fn modifier_display(tag: &str) -> &'static str {
+    match tag {
+        "cmd" => {
+            #[cfg(target_os = "macos")]
+            {
+                "Cmd"
+            }
+            #[cfg(target_os = "windows")]
+            {
+                "Ctrl"
+            }
+            #[cfg(all(not(target_os = "macos"), not(target_os = "windows")))]
+            {
+                "Super"
+            }
+        }
+        "ctrl" => "Ctrl",
+        "alt" => {
+            #[cfg(target_os = "macos")]
+            {
+                "Option"
+            }
+            #[cfg(not(target_os = "macos"))]
+            {
+                "Alt"
+            }
+        }
+        "shift" => "Shift",
+        "super" => "Super",
+        _ => "",
+    }
+}
+
+pub(super) fn display_primary(primary: &str) -> String {
+    let trimmed = primary.trim();
+    if trimmed.is_empty() {
+        return "?".to_string();
+    }
+    // 单个字母键归一为大写显示（"a" → "A"）；其余原样（如 ";"、"F1"）。
+    if trimmed.chars().count() == 1 {
+        let ch = trimmed.chars().next().unwrap();
+        if ch.is_ascii_alphabetic() {
+            return ch.to_ascii_uppercase().to_string();
+        }
+    }
+    trimmed.to_string()
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub enum HotkeyTrigger {
+    RightOption,
+    LeftOption,
+    RightControl,
+    LeftControl,
+    RightCommand,
+    Fn,
+    RightAlt, // Windows synonym for RightOption
+    MediaPlayPause,
+    Custom,
+}
+
+impl HotkeyTrigger {
+    pub fn display_name(&self) -> &'static str {
+        match self {
+            HotkeyTrigger::RightOption => "右 Option",
+            HotkeyTrigger::LeftOption => "左 Option",
+            HotkeyTrigger::RightControl => "右 Control",
+            HotkeyTrigger::LeftControl => "左 Control",
+            HotkeyTrigger::RightCommand => "右 Command",
+            HotkeyTrigger::Fn => "Fn (地球键)",
+            HotkeyTrigger::RightAlt => "右 Alt",
+            HotkeyTrigger::MediaPlayPause => "⏯ Media 播放/暂停",
+            HotkeyTrigger::Custom => "自定义组合键",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub enum HotkeyMode {
+    Toggle,
+    Hold,
+    DoubleClick,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub enum HotkeyAdapterKind {
+    MacEventTap,
+    WindowsLowLevel,
+    Fcitx5,
+    /// Mobile platforms do not expose desktop global hotkey adapters.
+    Unavailable,
+}
+
+impl HotkeyAdapterKind {
+    pub fn display_name(&self) -> &'static str {
+        match self {
+            HotkeyAdapterKind::MacEventTap => "macOS Event Tap",
+            HotkeyAdapterKind::WindowsLowLevel => "Windows 低层键盘 hook",
+            HotkeyAdapterKind::Fcitx5 => "fcitx5 输入法插件",
+            HotkeyAdapterKind::Unavailable => "不可用",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct HotkeyKey {
+    pub code: String,
+}
+
+impl HotkeyKey {
+    pub fn new(code: impl Into<String>) -> Self {
+        Self { code: code.into() }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(default, rename_all = "camelCase")]
+pub struct HotkeyBinding {
+    pub trigger: HotkeyTrigger,
+    pub mode: HotkeyMode,
+    pub keys: Option<Vec<HotkeyKey>>,
+}
+
+impl HotkeyBinding {
+    pub fn effective_codes(&self) -> Vec<String> {
+        let Some(keys) = &self.keys else {
+            let code = legacy_trigger_code(self.trigger);
+            return if code.is_empty() {
+                Vec::new()
+            } else {
+                vec![code.to_string()]
+            };
+        };
+        keys.iter()
+            .map(|key| key.code.trim().to_string())
+            .filter(|code| !code.is_empty())
+            .collect()
+    }
+
+    pub fn display_label(&self) -> String {
+        let codes = self.effective_codes();
+        if codes.is_empty() {
+            return "未设置".to_string();
+        }
+        codes
+            .iter()
+            .map(|code| display_hotkey_code(code))
+            .collect::<Vec<_>>()
+            .join("+")
+    }
+}
+
+pub(super) fn legacy_trigger_code(trigger: HotkeyTrigger) -> &'static str {
+    match trigger {
+        HotkeyTrigger::RightOption | HotkeyTrigger::RightAlt => "AltRight",
+        HotkeyTrigger::LeftOption => "AltLeft",
+        HotkeyTrigger::RightControl => "ControlRight",
+        HotkeyTrigger::LeftControl => "ControlLeft",
+        HotkeyTrigger::RightCommand => "MetaRight",
+        #[cfg(target_os = "windows")]
+        HotkeyTrigger::Fn => "ControlRight",
+        #[cfg(not(target_os = "windows"))]
+        HotkeyTrigger::Fn => "Fn",
+        HotkeyTrigger::MediaPlayPause => "MediaPlayPause",
+        HotkeyTrigger::Custom => "",
+    }
+}
+
+pub(super) fn display_hotkey_code(code: &str) -> String {
+    let label = match code {
+        "ControlLeft" => "左Ctrl",
+        "ControlRight" => "右 Control",
+        "AltLeft" => "左Alt",
+        "AltRight" => "右Alt",
+        "ShiftLeft" => "左Shift",
+        "ShiftRight" => "右Shift",
+        "MetaLeft" | "OSLeft" => "左Win",
+        "MetaRight" | "OSRight" => "右Win",
+        "Fn" => "Fn",
+        "FnLock" => "FnLock",
+        "CapsLock" => "CapsLock",
+        "ScrollLock" => "ScrLock",
+        "Pause" => "Pause",
+        "PrintScreen" => "PrtSc",
+        "Backspace" => "Backspace",
+        "Tab" => "Tab",
+        "Enter" => "Enter",
+        "Space" => "Space",
+        "Insert" => "Insert",
+        "Delete" => "Delete",
+        "Home" => "Home",
+        "End" => "End",
+        "PageUp" => "PageUp",
+        "PageDown" => "PageDown",
+        "ArrowUp" => "Up",
+        "ArrowDown" => "Down",
+        "ArrowLeft" => "Left",
+        "ArrowRight" => "Right",
+        "NumpadAdd" => "Num+",
+        "NumpadSubtract" => "Num-",
+        "NumpadMultiply" => "Num*",
+        "NumpadDivide" => "Num/",
+        "NumpadDecimal" => "Num.",
+        "NumpadEnter" => "NumEnter",
+        "Mouse4" => "Mouse4",
+        "Mouse5" => "Mouse5",
+        "Backquote" => "`",
+        "Minus" => "-",
+        "Equal" => "=",
+        "BracketLeft" => "[",
+        "BracketRight" => "]",
+        "Backslash" => "\\",
+        "Semicolon" => ";",
+        "Quote" => "'",
+        "Comma" => ",",
+        "Period" => ".",
+        "Slash" => "/",
+        _ => "",
+    };
+    if !label.is_empty() {
+        return label.to_string();
+    }
+    if let Some(letter) = code.strip_prefix("Key") {
+        if letter.len() == 1 {
+            return letter.to_string();
+        }
+    }
+    if let Some(digit) = code.strip_prefix("Digit") {
+        if digit.len() == 1 {
+            return digit.to_string();
+        }
+    }
+    if let Some(num) = code.strip_prefix("Numpad") {
+        if num.len() == 1 && num.as_bytes()[0].is_ascii_digit() {
+            return format!("Num{num}");
+        }
+    }
+    code.to_string()
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct HotkeyCapability {
+    pub adapter: HotkeyAdapterKind,
+    pub available_triggers: Vec<HotkeyTrigger>,
+    pub requires_accessibility_permission: bool,
+    pub supports_modifier_only_trigger: bool,
+    pub supports_side_specific_modifiers: bool,
+    pub explicit_fallback_available: bool,
+    pub status_hint: Option<String>,
+}
+
+impl HotkeyCapability {
+    pub fn current() -> Self {
+        #[cfg(mobile)]
+        {
+            return Self {
+                adapter: HotkeyAdapterKind::Unavailable,
+                available_triggers: Vec::new(),
+                requires_accessibility_permission: false,
+                supports_modifier_only_trigger: false,
+                supports_side_specific_modifiers: false,
+                explicit_fallback_available: false,
+                status_hint: Some(
+                    "移动端不支持全局热键；请使用应用内录音按钮或悬浮窗（需授权）。".into(),
+                ),
+            };
+        }
+
+        #[cfg(target_os = "macos")]
+        {
+            Self {
+                adapter: HotkeyAdapterKind::MacEventTap,
+                available_triggers: vec![
+                    HotkeyTrigger::RightOption,
+                    HotkeyTrigger::LeftOption,
+                    HotkeyTrigger::RightControl,
+                    HotkeyTrigger::LeftControl,
+                    HotkeyTrigger::RightCommand,
+                    HotkeyTrigger::Fn,
+                    HotkeyTrigger::Custom,
+                ],
+                requires_accessibility_permission: true,
+                supports_modifier_only_trigger: true,
+                supports_side_specific_modifiers: true,
+                explicit_fallback_available: false,
+                status_hint: Some("授权辅助功能后，通常需要完全退出并重新打开 OpenLess。".into()),
+            }
+        }
+
+        #[cfg(target_os = "windows")]
+        {
+            return Self {
+                adapter: HotkeyAdapterKind::WindowsLowLevel,
+                available_triggers: vec![
+                    HotkeyTrigger::RightControl,
+                    HotkeyTrigger::RightAlt,
+                    HotkeyTrigger::LeftControl,
+                    HotkeyTrigger::RightCommand,
+                    HotkeyTrigger::MediaPlayPause,
+                    HotkeyTrigger::Custom,
+                ],
+                requires_accessibility_permission: false,
+                supports_modifier_only_trigger: true,
+                supports_side_specific_modifiers: true,
+                explicit_fallback_available: false,
+                status_hint: Some(
+                    "默认建议使用“右Ctrl + 单击”；若更习惯按住说话，可在录音设置里切回“按住”。若无响应，可在权限页查看 hook 安装状态。"
+                        .into(),
+                ),
+            };
+        }
+
+        #[cfg(all(not(target_os = "macos"), not(target_os = "windows"), not(mobile)))]
+        {
+            Self {
+                adapter: HotkeyAdapterKind::Fcitx5,
+                available_triggers: vec![
+                    HotkeyTrigger::RightAlt,
+                    HotkeyTrigger::RightControl,
+                    HotkeyTrigger::LeftControl,
+                    HotkeyTrigger::Custom,
+                ],
+                requires_accessibility_permission: false,
+                supports_modifier_only_trigger: true,
+                supports_side_specific_modifiers: true,
+                explicit_fallback_available: false,
+                status_hint: Some(
+                    "Linux 使用 fcitx5 插件监听热键和提交文字；无需桌面环境额外配置。".into(),
+                ),
+            }
+        }
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct HotkeyInstallError {
+    pub code: String,
+    pub message: String,
+}
+
+impl std::fmt::Display for HotkeyInstallError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{} ({})", self.message, self.code)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct HotkeyStatus {
+    pub adapter: HotkeyAdapterKind,
+    pub state: HotkeyStatusState,
+    pub message: Option<String>,
+    pub last_error: Option<HotkeyInstallError>,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub enum WindowsImeInstallState {
+    Installed,
+    NotInstalled,
+    RegistrationBroken,
+    NotWindows,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct WindowsImeStatus {
+    pub state: WindowsImeInstallState,
+    pub using_tsf_backend: bool,
+    pub message: String,
+    pub dll_path: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct PlatformCapabilities {
+    pub platform: String,
+    pub supports_ime_input: bool,
+    pub supports_overlay: bool,
+    pub supports_desktop_hotkey: bool,
+    pub supports_tray: bool,
+    pub supports_local_asr: bool,
+    pub supports_in_app_dictation: bool,
+    pub supports_auto_update: bool,
+}
+
+impl PlatformCapabilities {
+    pub fn current() -> Self {
+        #[cfg(target_os = "android")]
+        {
+            Self {
+                platform: "android".to_string(),
+                supports_ime_input: false,
+                supports_overlay: true,
+                supports_desktop_hotkey: false,
+                supports_tray: false,
+                supports_local_asr: false,
+                supports_in_app_dictation: true,
+                supports_auto_update: true,
+            }
+        }
+
+        #[cfg(all(
+            any(target_os = "android", target_os = "ios"),
+            not(target_os = "android")
+        ))]
+        {
+            Self {
+                platform: "mobile".to_string(),
+                supports_ime_input: false,
+                supports_overlay: false,
+                supports_desktop_hotkey: false,
+                supports_tray: false,
+                supports_local_asr: false,
+                supports_in_app_dictation: false,
+                supports_auto_update: false,
+            }
+        }
+
+        #[cfg(not(any(target_os = "android", target_os = "ios")))]
+        {
+            Self {
+                platform: "desktop".to_string(),
+                supports_ime_input: cfg!(target_os = "windows"),
+                supports_overlay: true,
+                supports_desktop_hotkey: true,
+                supports_tray: true,
+                supports_local_asr: cfg!(any(target_os = "macos", target_os = "windows")),
+                supports_in_app_dictation: false,
+                supports_auto_update: true,
+            }
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub enum HotkeyStatusState {
+    Starting,
+    Installed,
+    Failed,
+}
+
+impl Default for HotkeyStatus {
+    fn default() -> Self {
+        Self {
+            adapter: HotkeyCapability::current().adapter,
+            state: HotkeyStatusState::Starting,
+            message: Some("正在安装全局快捷键监听".into()),
+            last_error: None,
+        }
+    }
+}
+
+impl Default for HotkeyBinding {
+    fn default() -> Self {
+        // 注意：keys 必须是 None，不能预填具体 code。
+        //
+        // 原因：HotkeyBinding 用 `#[serde(default)]` **结构级 default**——反序列化时
+        // 整个 struct 先按 Default 填充再让 JSON 字段覆盖。如果这里 keys 预填了
+        // Some([...])，那么旧 prefs 里只写 `{"trigger":"rightControl","mode":"toggle"}`
+        // （不带 keys 字段）会被反序列化成 `{trigger=RightControl, keys=Some([默认值])}`
+        // 即 trigger 跟 keys 完全不一致——effective_codes() 直接信任 keys，导致
+        // 实际生效的快捷键跟用户当年选的 trigger 对不上。
+        // 现在 keys=None 时 effective_codes() 走 legacy_trigger_code(trigger) 路径，
+        // 跟 trigger 自动同步。
+        #[cfg(target_os = "windows")]
+        {
+            Self {
+                trigger: HotkeyTrigger::RightControl,
+                mode: HotkeyMode::Toggle,
+                keys: None,
+            }
+        }
+
+        #[cfg(not(target_os = "windows"))]
+        {
+            Self {
+                trigger: HotkeyTrigger::RightOption,
+                mode: HotkeyMode::Toggle,
+                keys: None,
+            }
+        }
+    }
+}
